Rename package, move intern specific modules to intern directory
