+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@dojo/intern-helper",
+  "name": "@dojo/test-extras",
   "version": "2.0.0-pre",
   "description": "A set of modules to help with testing Dojo 2 with Intern",
   "engines": {
@@ -9,7 +9,7 @@
   "private": true,
   "homepage": "http://dojotoolkit.org",
   "bugs": {
-    "url": "https://github.com/dojo/intern-helper/issues"
+    "url": "https://github.com/dojo/test-extras/issues"
   },
   "license": "BSD-3-Clause",
   "main": "main.js",
@@ -20,7 +20,7 @@
   ],
   "repository": {
     "type": "git",
-    "url": "https://github.com/dojo/intern-helper.git"
+    "url": "https://github.com/dojo/test-extras.git"
   },
   "scripts": {
     "prepublish": "grunt peerDepInstall",
