+++ src/intern/ClientErrorCollector.ts
