+++ tests/functional/intern/ClientErrorCollector.html
