+++ tests/unit/all.ts
@@ -1,3 +1,3 @@
-import './ClientErrorCollector';
 import './main';
+import './intern/all';
 import './support/all';
