+++ tests/unit/intern/all.ts
@@ -1,0 +1,1 @@
+import './ClientErrorCollector';
